Clean-up sticky slider
