+++ src/assets/toolkit/styles/toolkit.scss
@@ -120,7 +120,7 @@
 @import 'components/logo';
 @import 'components/address';
 @import 'components/accordion';
-@import 'components/stickyslider';
+@import 'components/sticky-slider';
 @import 'components/quote';
 @import 'components/article';
 @import 'components/article-card';
