+++ src/views/example-pages/home.html
@@ -19,7 +19,7 @@ title: Home
 
 <section class="section">
   <div class="section__body">
-    {{>stickyslider}}
+    {{>sticky-slider}}
   </div>
 </section>
